requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
requests
requests-oauthlib
python-dotenv
gunicorn
google.generativeai